fix(tests): Increase attestation timeout and update Makefile
The `test_attestation` test may time out while waiting for SSH access,
as VM boot can exceed the previous 300s limit. Increase the maximum
wait time to 600s to prevent this.

Also, update the `integration-tests` target in the Makefile to add
the `crds-rs` dependency and ensure the namespace is created before
applying the manifests. And fix the race condition in the tests.

Signed-off-by: Yalan Zhang <yalzhang@redhat.com>